+++ code/sound/soundfx/soundfxgameplaylogic.cpp
@@ -392,7 +392,7 @@ void SoundFXGameplayLogic::HandleEvent( EventEnum id, void* pEventData )
             break;
 
         case EVENT_FE_CREDITS_NEW_LINE:
-            playCreditLine( reinterpret_cast<int>(pEventData) );
+            playCreditLine( reinterpret_cast<intptr_t>(pEventData) );
             break;
 
         default:
